Add initial_mrkdwn and replace_assets options. (#9)
* Add initial_mrkdwn and replace_assets options.

These are the last two features I need for this action to be perfect
for my current continuous deployment flow! I want to be able to
specify markdown to use as a template when creating the initial
release, but then be able to hand-edit it later, and be sure that my
edits will not be destroyed by future action runs. This new option is
a backwards-compatible way of achieving that. The old `body_mrkdwn`
option works the same as it always has, but you can now specify an
`initial_mrkdwn` value that is used only when creating a new release.
If you don't specify `initial_mrkdwn`, then `body_mrkdwn` is used
instead for the new release, and if neither was given, your standard
fallback string is still used.

But if you specify only `initial_mrkdwn` and no `body_mrkdwn`, then
the markdown text is only used for creating new releases, and existing
release descriptions are left unmodified.

The second argument, `replace_assets`, controls whether to do the new
feature you added this weekend. My jobs run in two different modes,
because they are based on the Maven package management system which is
the foundation of the Java world. Snapshot releases are works in
progress, and can be released many times. Their assets get updated
with each release. But a non-snapshot release is final, and its
artifacts _never_ change. So I want to replace assets when building
snapshot releases, but flag it as an error if anyone ever attemts to
replace assets for a non-snapshot release. And my jobs can determine
the kind of release from the most recent git tag.

* Remove redundant text

Sorry, I was up too late!

Co-Authored-By: Xotl <edgardo333@gmail.com>

* Fix input name

Nice catch, I should have reviewed this the next morning.

Co-Authored-By: Xotl <edgardo333@gmail.com>

Co-authored-by: Xotl <edgardo333@gmail.com>