Also fix warnings from deprecated core actions.
You may wish to update @actions/github, which are up to version
5.1.1 (you are specifying ^2.0.0 and receiving 2.2.0). But this is
such a large version change I did not want to assume it would work.

Your specification for node-modules/mime-types of ^2.1.25 is already
receiving the latest, ^2.1.35, so that is less urgent to update, but
you may want to as well.