Fix strict settings
It looks like these scripts were trying to use strict Javascript
parsing, but had a typo in the directive. (Or maybe this is a
different directive that I don't know about, I admit I do very little
Javascript development. If that is the case, please just ignore this
pull request. But I could not find any suggestion that this would do
anything.)

Also, `index.js` was missing the directive entirely, and I though you
might want it there as well?